Implement surrounding_trivia
