+++ tests/node.rs
@@ -1,21 +1,17 @@
 use full_moon::{node::Node, parse};
 
 #[test]
-fn surrounding_ignore_tokens() {
+fn surrounding_trivia() {
     let ast = parse(include_str!("cases/pass/local-assignment-5/source.lua")).unwrap();
     let stmt = ast.nodes().iter_stmts().nth(1);
 
-    let (prev, next) = stmt.surrounding_ignore_tokens(&ast).unwrap();
+    let (prev, _) = stmt.surrounding_trivia();
 
     let mut prev = prev.into_iter();
     assert_eq!(prev.next().unwrap().to_string(), "\n");
     assert_eq!(prev.next().unwrap().to_string(), "-- Then a comment");
     assert_eq!(prev.next().unwrap().to_string(), "\n");
     assert_eq!(prev.next(), None);
-
-    let mut next = next.into_iter();
-    assert_eq!(next.next().unwrap().to_string(), "\n");
-    assert_eq!(next.next(), None);
 }
 
 #[test]
@@ -26,4 +22,18 @@ fn test_similar() {
     assert!(stmts[0].similar(stmts[1]));
     assert!(stmts[1].similar(stmts[0]));
     assert!(!stmts[0].similar(stmts[2]));
+}
+
+#[test]
+fn test_tokens_collect() {
+    let source = parse("local abcd = 1").unwrap();
+    let tokens = source.nodes().tokens().collect::<Vec<_>>();
+    assert_eq!(tokens.len(), 4);
+}
+
+#[test]
+fn test_tokens_back() {
+    let source = parse("local abcd = 1").unwrap();
+    let mut tokens = source.nodes().tokens();
+    assert_eq!(tokens.next_back().unwrap().to_string(), "1");
 }
